Move the phone's transport, pairing and session state out of main
main.rs held the remote API, the pairing link parser, the room projection and
every screen in one 2227-line file, so nothing in it could be found by the name
of what it did. The three clusters that are not UI move to siblings: api.rs
speaks the shared protocol over QUIC, pairing.rs owns the link the desktop hands
out, session.rs owns what the phone is attached to and how a room's events fold
into what it draws. Their tests move with them.

Pairing's three free functions become Credentials methods, since each of them
took a pairing and returned something derived from it. MobileRoomProjection
keeps its fields private and answers event_count instead, which is all the
transcript's scroll effect wanted.

main.rs is now the launch glue and the screens: 1400 lines, still too many, but
the remainder is one component rather than four unrelated concerns.
